requirements-dev.txt
build
copier>=9,<10
pytest>=8
PyYAML>=6

